+++ cat-core/src/main/java/com/dianping/cat/report/service/AbstractReportService.java
@@ -174,7 +174,7 @@ public abstract class AbstractReportService<T> implements ReportService<T> {
 		try {
 			m_monthlyReportDao.insert(report);
 
-			int id = report.getId();
+			long id = report.getId();
 			MonthlyReportContent proto = m_monthlyReportContentDao.createLocal();
 
 			proto.setReportId(id);
