+++ cat-home/src/main/java/com/dianping/cat/report/page/transaction/service/TransactionReportService.java
@@ -137,7 +137,7 @@ public class TransactionReportService extends AbstractReportService<TransactionR
 		}
 	}
 
-	private TransactionReport queryFromMonthlyBinary(int id, String domain) {
+	private TransactionReport queryFromMonthlyBinary(long id, String domain) {
 		MonthlyReportContent content = m_monthlyReportContentDao.findByPK(id);
 
 		if (content != null) {
